Doctests: Respect fbcode default link mode
Summary:
This adds a `doctest_link_style` attribute to the `rust_library` rules, which matches the `link_style` on Rust binaries and Rust unittests.

This is probably not the greatest solution to this problem. Ideally, we would just make Rust doctests be a rule like the unittests, in which case we no longer have to do this. I'm not going to do that here though.

Reviewed By: dtolnay

Differential Revision: D45073247

fbshipit-source-id: fa4a5f2cfc93702f01a36bf8c3baea283c80a90c