Cut strip_prefix from the short_path observable by downstream targets
Summary:
My `strip_prefix` implementation from D45075567 ended up not doing quite what I wanted.

In the case of the following pair of targets:

```
http_archive(
    name = "bitflags-1.3.2.crate",
    urls = ["https://crates.io/api/v1/crates/bitflags/1.3.2/download"],
    sha256 = "bef38d45163c2f1dde094a7dfd33ccf595c92905c8f8f4fdc18d06fb1037718a",
    strip_prefix = "bitflags-1.3.2",
)

rust_library(
    name = "bitflags-1.3.2",
    srcs = [":bitflags-1.3.2.crate"],
    crate = "bitflags",
    crate_root = "bitflags-1.3.2.crate/src/lib.rs",
    edition = "2018",
    rustc_flags = ["--cap-lints=allow"],
)
```

the user's expectation is that the `http_archive` represents a directory called "bitflags-1.3.2.crate" which contains a file called "src/lib.rs". These are the only two things about the http_archive which are relevant to the downstream rust_library target shown after it. We could replace the `http_archive` with a 100% equivalent artifact created using `filegroup` with name "bitflags-1.3.2.crate" containing a file "src/lib.rs" and nothing downstream should notice the difference.

The use of `strip_prefix` is meant to be a private implementation detail of that target. It controls how the target translates its tgz into a directory containing a file called "src/lib.rs". The value of `strip_prefix` should never make a difference to anything downstream. We could produce a 100% equivalent artifact by changing the URL to a different tgz which removes a layer of nesting, and removing the `strip_prefix`, and nothing downstream should notice the difference.

This interpretation is consistent with Bazel's `http_archive`'s implementation of `strip_prefix`.

Unfortunately it's not what I implemented in D45075567. As implemented, a target with `http_archive(name = "a", strip_prefix = "b/c", ...)` containing file "d" would appear to downstream targets like `rust_library` as a directory named "b/c" containing file "d", *not* a directory named "a" containing file "d" as intended.

Reviewed By: krallin

Differential Revision: D45097536

fbshipit-source-id: 98abdbf26f07b72020226762ee90968059b7906e