Remove link_postprocessing
Summary:
This was ever only used for focused_debugging, which to properly support, we've had to make a first party integration vs just a post process step.

As such, this field is now no longer used, and it's use creates complications (with cache hits, etc). Additionally, the API should likely be plural, allowing multiple post processors.

Regardless, remove support as until there is the need for it, keeping this API just adds additional complexity.

Reviewed By: blackm00n

Differential Revision: D45130217

fbshipit-source-id: 9ed98f70787229a85b21689d3fd2f50a1ed5ff01