Fix up some LSP bugs
Summary: Fixing these because otherwise it breaks my future diffs. LSP attempts to build an index of global functions. It previously got that wrong by skipping functions inside modules (aka almost all of them) and pretending types were modules (which is true for providers, which happened to be what it tested). Switch to properly selecting the globals.

Reviewed By: bobyangyf

Differential Revision: D45197534

fbshipit-source-id: e99db350dd8b23e89bb68e44dbb683821e83ac2f