buck2: active commands: show the same span counts as the console
Summary:
In particular, the counts should reflect roots, and not all spans.

The long term plan is for this to actually *be* the console, so that seems like
something we'd want. For the short term this also makes it a bit more useful
for IDE folks wanting to start building upon this.

Reviewed By: milend

Differential Revision: D45181686

fbshipit-source-id: 78eac9d4d93be851d618da0ec10fb67f9efe142b