offline builds: Add repository metadata
Summary:
Adds repository metadata so we can incorporate this (especially repo name) into the resulting offline archive.

The main idea is that if a manifest comes from fbsource, we'll be able to rehydrate files into an "fbsource" directory inside the archive at the revision upon which the manifest was generated.

Reviewed By: krallin

Differential Revision: D44824382

fbshipit-source-id: eacf4d8b85397db05eda98c84ae7659acc9ed78e