Focused -> Selective Part I
Summary:
"Focused debugging" is derived from buck1's "focus" behavior of which targets were included in the Xcode workspace.

Buck2's project generation allows more toggles to select what is included in the workspace, and on demand integrations do the same.

Rename "focused debugging" to "selective debugging" as we more formalize the integration into the rules and other systems.

In this diff, we update the prelude, toolchain and downstream references.

Up the stack, we'll rename others.

Reviewed By: milend

Differential Revision: D45185740

fbshipit-source-id: faf4cc94152af633a89afb08899baae12d6920a7