Add a wrapper around AnalysisActions
Summary:
This is for eventually supporting execution platforms.

End state is to have some wrapper around the actions factory, and provide some accessors for users to get the resolved execution platforms if requested to use in their builds.

In the long run, we want to be able to support multiple calls to action_factory, such that each time we can create different execution platforms (this is execution groups, which is not implemented). But, let's start with a simpler implementation for now which will support a single execution platform.

Reviewed By: bobyangyf

Differential Revision: D44992578

fbshipit-source-id: 87f00ce63c049fb5d7f6f3850cf3aada75eaf5ae