buck2: clean stale: don't invalidate declared paths
Summary:
There are two codepaths here that can end up invalidating declared artifacts.
This is just as bad as deleting an active artifact since it means that artifact
will not get materialized again.

This diff redoes this code to minimize the number of conditionals and have
*one* place where we enqueue something to be invalidated,

This loses a bit of functionality compared to the previous state (which is when
the aforementioned bugs got introduced). Namely, if we delete all the entries
under a path, we don't delete that path.

I'm going to argue this is fine because we'll pick them up the next time you
run `buck2 clean --stale`, which seems *good enough*. I updated the tests
accordingly.

This also doesnt handle non-UTF8 paths on disk, but that actually wasn't
supported by the original implementation, either (though it might have seemed
like it was, it doesn't have the ability to *actually delete something that
isn't a UTF8 path*, so it's rather obvious that this is broken).

 ---

Note that the actual issue being fixed here seems esoteric but actually
happens rather often enough. Specifically, if you got hybrid execution,
but RE wins the race, then we'll create the parent dirs for an output
that we're not downloading immediately (because they are on RE).

If the parent dir turns out to have *no* materialized outputs at all
in it (e.g. `__objects__`), then the current logic will delete it along
with everything under it from disk, but **also from the materializer state**,
which means we'll remove declared outputs.

if we then try to access those outputs later, we will not materialize them,
and we will get an error.

See this doc for a specific instance of this + debugging:

https://docs.google.com/document/d/12jushxNanKjTAbXl2t0nafIFiX8YUGJD-Zg3Jc5t-0A/edit

Reviewed By: christolliday

Differential Revision: D45231964

fbshipit-source-id: ac7ea009a76a6f76c403bb150dc40740c284225e