buck2: build: only accumulate results at one level
Summary:
Right now we have 3 different levels that collect futures. This is problematic
because I would like to implement support for a `--fail-fast` flag, and doing
that is quite challenging if all the levels need to be aware of it.

So, in order to simplify this, actually move the aggregation to just one level.
To do this we update everything to just return a stream of results and now we
we have a single collect.

Note that while this is a `try_collect`, that doesn't implement fail-fast yet:
because the `Err` variants here are when load or analysis fails (in which case
we do fail fast already, this may be a bit weird), but not when build fails (in
which case we do not: we return a Ok variant that indicates it failed).

Reviewed By: stepancheg

Differential Revision: D43693052

fbshipit-source-id: bac109890f896d973311d7c8a9693f3650f63521