buck2: build: accumulate results independently for each output
Summary:
Like it says in the title. If you've got lots of artifacts for a given target,
then to implement fail-fast properly you want to stop on the first failure, and
not wait until all results for that target are ready.

This implements that by producing a stream of messages for each target being
built, which we later collect to reconstruct the original BuildTargetResult.

Reviewed By: stepancheg

Differential Revision: D43694047

fbshipit-source-id: 1672e31c9d6997685460962b6bbed27d20b38cf0