Fix execution directory for the test shell
Summary:
In the buck2 shell test, the execution directory was set at the root of the buck project, leading tests to write into root/log, which leads to an "isolation break", which leads to tests having non reproducible behaviour.

We set up the execution directory to be the temporary directory created where the logs and home would be placed, just as with buck2 test.

Reviewed By: TheGeorge

Differential Revision: D45157924

fbshipit-source-id: 4b0a9d3f3626296056a62b89f7fce0350d6ef961