Spawn package loading only once
Summary:
When performing command like:

```
buck2 targets //foo:bar //foo:baz
```

we only need to load package `//foo` once.

Before this diff package was loaded once, but the task to load it was created twice.

This diff deduplicates these tasks.

Context: RL TD takes target list from BTD and then filters targets with `ctargets`. So naturally there is a lot of duplicated packages in the input. This change makes evaluation a bit cheaper. Probably not significantly, but this is not too complicated change.

Reviewed By: bobyangyf

Differential Revision: D45283163

fbshipit-source-id: 3d40417bc05334b32a9d026c6419892fa1dfdad2