Fix proc macro doctests for rust 1.69+
Summary:
There is a change in how `rustdoc` of Rust 1.68 and older vs 1.69 (and presumably newer) handle doc tests for proc macro crates.

Previously we were running `rustdoc --test --crate-type=bin src/lib.rs`, because `bin` is the crate type that the tests get built in, and this did the right thing for building the right flavor of all the dependencies.

In 1.69.0 this no long works; it wants `rustdoc --test --crate-type=proc-macro src/lib.rs` or else the doc test will fail with:

```
error: the `#[proc_macro_derive]` attribute is only usable with crates of the `proc-macro` crate type
  --> /data/sandcastle/boxes/eden-trunk-hg-fbcode-fbsource/buck-out/v2/gen/fbcode/74eb442a07b49473/thrift/lib/rust/test/transitive/___get_struct_name_derive__/__srcs/macro.rs:20:1
   |
20 | #[proc_macro_derive(GetStructName)]
   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

error: aborting due to previous error

✗ Fail: thrift/lib/rust/test/transitive:_get_struct_name_derive - main (0.0s)
```

I deferred fixing this from the fbcode toolchain update in {D45267362}.

Reviewed By: diliop

Differential Revision: D45292128

fbshipit-source-id: c704e61dcf43ad107e2be67ab55af712b6301a17