buck2: fix detection of buck-out not being redirected
Summary:
This doesn't actually seem to work at all right now: I have an example from
IanChilds where I know for a fact his redirection went away (because he was missing
all his files and `eden doctor` reported the redirection was busted), I know
he was using a version of Buck2 that had this, and yet it didn't report
anything. This also repros for myself (see test plan).

I don't think we actually care about buck-out being on FUSE or NFS.
Specifically, we care about buck-out not being on the EdenFS mount when EdenFS
itself is in use.

So, just implement that.

Reviewed By: get9

Differential Revision: D45308806

fbshipit-source-id: ef5d99846e35d37d65bf4bd34224eba5a8e387e0