offline builds: Materialize download_file outputs to (and read from) separate "offline" cache
Summary:
One additional problem with including network action outputs (e.g. `ctx.action.download_file` in rules) is that, regardless of whether the output is stored or not, we always at least do a HEAD request to determine if the remote file has changed.

This won't work for offline environments, so we need to additionally cache this file in a separate place. In this diff, a new buck-out subdir is introduced: the offline cache (buck-out/v2/offline-cache).

This cache will store cached action outputs such that, when action paths exist for a given action, we'll simply treat download_file actions as copies and materialize this output into the proper buck-out location from the action output cache.

Next diff does the same for cas_artifact.

Reviewed By: krallin

Differential Revision: D45057573

fbshipit-source-id: 9b791cde574bcbf6bea477a4931c561b2735b9d0