offline builds: Retain absolute paths from config parsing
Summary:
Prior to this diff we were explicitly filtering out non-absolute paths with the thought that we didn't want to include these external configs in the final archive.

Well spoilers, but we _do_ actually need these to pick up configs like `host_config.gvfs` and some other ones that cause fbcode builds to have the wrong configuration if missing.

Reviewed By: krallin

Differential Revision: D45233443

fbshipit-source-id: 5dc744281daf4bd7e985757770766ad01d9e87d3