offline builds: Materialize cas_artifact outputs to (and read from) separate "offline" cache
Summary:
As previously, do the same thing we did with `download_file` with `cas_artifact`. When `--config buck2.use_network_action_offline_cache` is passed during a build, we'll attempt to read CAS artifact fetch outputs from the local offline action cache _first_ before then reaching out to the network.

This is done so that we can do fully offline builds without having to reach out to the network to e.g. determine if we need to re-download the cas blob.

Reviewed By: krallin

Differential Revision: D45057574

fbshipit-source-id: d86335aa60d42f61cc5b266854182250496417b5