offline builds: Leverage buck to drive file/directory/symlink partitioning
Summary:
When reading paths, buck does a lot of the hard work of figuring out whether:
* A path is a file or directory
* A path is a symlink or traverses one
* If it's a symlink, breaking it apart into the requisite symlink + remaining path components

Rather than re-implement all of this, why don't we leverage buck to tell the offline build archiver tool what kinds of paths each one is.

This alters the manifest definition to include a few new categories:
* **Paths** are normal repo-relative files/dirs
* **Relative symlinks** are repo-relative symlinks with targets that are also repo-relative (e.g. some link pointing to something else inside the repo)
* **External symlinks** are repo-relative symlinks with targets outside of the repo, e.g. gvfs links

Note that the way this is implemented introduces a perf hit during I/O tracing because we not only read each file / dir (from e.g. `read_path_if_exists`), but we also get the file metadata so we know whether and what kind of symlink this path is. This is the only way to be sure that we categorize repo entries effectively and greatly simplifies the archiving program.

It's acceptable (to me) to have double file accesses per file open if it means we can simplify the logic of the build archiver tool; getting this right is tricky and I would much prefer to lean on buck's work rather than reimplement it elsewhere.

Reviewed By: krallin

Differential Revision: D45172949

fbshipit-source-id: cbf0308dfd421aa136576f197c865e9f2e693ae9