offline builds: Add .buck2(-previous) dotslash configs in archiver, not buck2 itself
Summary:
We need to ensure the .buck2 and .buck2-previous files get included in the offline trace. buck2 itself used to do this, but this is suitably meta-only behavior (e.g. dotslash is still currently internal-only).

Note that because of the way repo entries are added to the repo - particularly directories - we'll already have these entries in the manifest because fbsource/ is read as a directory. The net result is that even by moving these to the offline archiver program, buck2 will _still_ include these in the manifest.

Still, since this is archival specific for internal use-cases, let's move these to the archiver program.

Reviewed By: krallin

Differential Revision: D45233442

fbshipit-source-id: a667baa96906e9f15bcaeac0d12640398598676d