buck2: download: claim the CommandExecutionManager later
Summary:
This will enable the next diff. The reason I want this is to allow injecting
validation other outputs before claiming the manager (at which point it cannot
be reused anymore), and potentially just returning the manager in that case.

There's obviously a small behavior change here, in that we'll download trees
out of RE before we claim, but that's a) rather rare and b) rather fast, so
it's not really worth worrying about .

Reviewed By: bobyangyf

Differential Revision: D45316351

fbshipit-source-id: ee02a6749e93abbdea2746ecffbcd3e32e017696