Improve "--no-remote-cache" setup
Summary:
Today "--no-remote-cache" option making buck to ignore the action cache checks but still tells RE workers to write to the action cache and which is likely not a goal at all when people are using this option and moreover could lead to some action cache mess due to some nondeterministic actions.

So this diff changes the behaviour and when only "--no-remote-cache" specified we disable action cache writes.
If we want to disable action cache reads but enable writes (let say because we want to override the action cache) then we must pass two flags: "--no-remote-cache --do-cache-remotely"

Reviewed By: krallin

Differential Revision: D45270212

fbshipit-source-id: cf42ff5c0f8a430fd4464b8348f9fde5e6c1a0fa