log: Rename `last` to `path`, add alias
Summary:
All of the other log subcommands are named after what they output. `buck2 log last` is the one exception to this, as it seems to be named after the recency of the log. This is pretty inconsistent imo, since `buck2 log show` also outputs the most recent log.

Alias should keep this from breaking anybody. Made the alias visible to avoid confusion, in case someone finds references to this in code/documentation

Reviewed By: bobyangyf

Differential Revision: D45380438

fbshipit-source-id: c4ff32d6651c0fa66c046a3fd47c5c6666ba2607