Drop `buck2_core` dep from `buck2_data`
Summary:
I'm currently experimenting with some tooling locally that requires reading an event log. I could use `buck2 log show` and then parse the json, but that's a bit of a headache. Since I'm already writing Rust, it would be nice if I could instead reuse the perfectly good Rust code in `buck2_cli_proto` that's already capable of reading an event log. The goal of this and the next diff is to improve the ergonomics of doing that. I don't know yet if I'm actually going to end up landing my code that uses this, but it seems like a reasonable thing to want on its own, anyway.

This diff improves things by inverting the dependency edge between `buck2_core` and `buck2_data`. This only requires moving some trait impls, and in exchange means I can depend on `buck2_cli_proto` without having to pull in all of `buck2_core`.

Reviewed By: stepancheg

Differential Revision: D45385712

fbshipit-source-id: 3cac4673cb9b5d69a2e0f78820fa02b06c81f5d0