Delete support for versioned_dirs=false
Summary:
The `buildscript_run` functionality exposed through the buck2 prelude already assumes you are using **either** `vendored = false` **or** `versioned_dirs = true`.

https://www.internalfb.com/code/fbsource/[3e95bc3bf154bb34b3228ccad21c11e59d1b596c]/fbcode/buck2/prelude/rust/cargo_buildscript.bzl?lines=110-111%2C113%2C115-121%2C124

While it would be possible to make this compatible with `versioned_dirs = false`, the generated target would need to be noisier and it's not clear there is demand for this mode.

Reviewed By: zertosh

Differential Revision: D45448073

fbshipit-source-id: 441292e608296820f58a12db1d9d44419cd8747d