buck2: don't report action cache corruption for declared digests
Summary:
Like it says in the title.

Updating just the body of the function here would be weird, so let's also
rename it to express what we actually want from this function, i.e. "does the
action cache guarantee this result exist or not?", which for declared outputs
is naturally false (whereas the previous name, i.e. "has this action result
expired" was more ambiguous).

Reviewed By: lmvasquezg

Differential Revision: D45474621

fbshipit-source-id: 630d4183a126ffb1780c29c39c4fa48dfc9c6ac6