Fix unaligned aaa_ptr() when cap == 0
Summary: D44476382 introduces the ability to compile rust stdlib from source, which adds additional safety checks in the form of debug-only assertions. This caught an issue with some unsafe rust in buck2 starlark, where `aaa_ptr()` was returning a misaligned pointer when `cap` == 0. The test case added triggers the issue.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D45450541

fbshipit-source-id: 7782b827b911644fa5727c5785f60bf033006ea3