Add a function is_root_cell
Summary: Seems odd to do `x == cells.root_cell()` when you might be able to do it cheaper. We currently can't do it cheaper, but one day it might be the case, so introduce the function. Also makes code slightly more readable, and to find the places a root cell is special.

Reviewed By: krallin

Differential Revision: D45477792

fbshipit-source-id: 90cc12841019717220f00c99f0758589259afb6d