execution: inherit standard windows envs
Summary:
Using OSS buck2 on Windows to run rust doctests fails because there is
no `TEMPDIR` set. On Windows, [`std::env::temp_dir`] will fallback to
`C:\Windows` when none of the common envs are set. To write to
`C:\Windows` (which is itself a bad idea), requires Administrator
privileges, so the tests fail.

Rather than just add the `TEMP` env that I need, I am enumerating all of
the standard built-in envs since this is pretty normal stuff.

On unix, this should probably cover `LANG`, `LC_*`, and `TZ`. Lots of
things depend on those envs.

[`std::env::temp_dir`]: https://doc.rust-lang.org/std/env/fn.temp_dir.html

Reviewed By: krallin

Differential Revision: D45556479

fbshipit-source-id: 52b79f3d30c9b389cc1f3473ac618eca4d238568