Add disable_cancellation to ExplicitlyCancellableFuture
Summary:
Add disable cancellation where we disable cancelling of the future from here on out.

This is slightly different from existing drop cancel behaviour, but DICE is the only user of this, and frankly it makes more sense for dice to forever disable cancellation than simply disable it while holding a guard.

Reviewed By: krallin

Differential Revision: D45419034

fbshipit-source-id: d37a2ecdea3fd2698622c40e98d7e0f3c5110457