Read fewer files in dex dynamic analysis
Summary:
We've observed that dynamic analysis for determining which files should go in each primary/secondary dex can be slow. Taking a look at it, almost all of the time goes into the `read_string()` call that we do on each file to get the primary/secondary class names and weight estimate: https://www.internalfb.com/code/fbsource/[85dbf5ae59ca]/fbcode/buck2/prelude/android/dex_rules.bzl?lines=569 There are > 30k such files for fb4a.

Instead, when we run "filter_dex", we can just produce a single file containing JSON for all of the dex libraries that it has filtered. We filter 100 libraries at a time, so this reduces the number of files that we have to read in dynamic analysis by 100x.

Reviewed By: cjhopman

Differential Revision: D45564377

fbshipit-source-id: 7342126c50355aed0ee8daf35eba4780c82fb29e