buck2: clean_output_paths: add a counter
Summary:
This doesn't use `fs_util::symlink_metadata` because a) we don't want to create
anyhow errors for error paths we expect and b) we want to match the error
(though admittedly that could be fixed with symlink_metadata_if_exists).

However, that also means it gets no stats... Fix it.

Reviewed By: IanChilds

Differential Revision: D45603410

fbshipit-source-id: 86d33b9d8412eef2980afad66cba7645af4a8c1d