Don't provide transitive debug data for static libraries
Summary:
When an `apple_library` with swift is linked, it's swiftmodule file must be included.

When an `apple_library` is linked as a link group, only the external debug info for whats in the link group should be included with that library.

However, as is, the logic of providing all transitive swift debug info breaks this requirement. Instead, we're propagating all debug info for swift modules that are transitively depended on, instead of the debug info for ***just*** this library.

We can fix this by providing the debug info for just this library, but then, we break the debug info in the case where the `apple_library` is linked shared and contains swift. For example: `fbsource//fbobjc/buck2/tests/swift:swift_dep_with_transitive_dep_via_objc_dep`

To fix this correctly, the MergedLinkInfo data propagation for shared links needs to take into account other related data to propagate. However, instead of tacking something on, the long term correct approach feels like we should use the linkable graph and use it for accumulating all ink related content. This is a much more extensive refactor, so as a quick workaround, we can check the type of linkage and pass the content we need.

This ensures that for a dylib with swift, we still accumulate the require debug info. And if an apple_library is linked with a link group, it will only get the debug info for whats being linked.

Reviewed By: milend

Differential Revision: D45569017

fbshipit-source-id: 36cec41781fd639b12615f208edc46ddb5bce94b