Refactor StarlarkOutputArtifact and ArtifactDeclaration
Summary:
This refactors StarlarkOutputArtifact to hold a StarlarkDeclaredArtifact
instead of holding the rust OutputArtifact. This then allows
us to change ArtifactDeclaration such that it always holds a
reference to the StarlarkDeclaredArtifact and doesn't need
different logic based on how it was declared.

In addition, this means we never try to convert a
non-declaredartifact into a declared one like we could before
(i.e. this deletes allocate_artifact_with_extended_associated_artifacts and
the corresponding thing is only supported on
StarlarkDeclaredArtifact directly)

Reviewed By: bobyangyf

Differential Revision: D44659206

fbshipit-source-id: 409399fec7ff7055d19a7d6e96dce7461f27a8e7