Log working set sizes
Summary: Currently, we don't log anything about memory on windows and that makes `buck2 status --snapshot` and our scuba datasets less useful for seeing what memory usage is like. This diff calls win32's `GetProcessMemoryInfo` to get the current and peak memory usage of a process and log that for buck2 so we can start tracking those info. Technically, on windows, these metrics are "working set size", not "resident set size", but I use the rss fields we already have in buck2 because (1) it's convenient and (2) windows definition of working set size sounds just like resident set size on unix

Reviewed By: krallin

Differential Revision: D45576309

fbshipit-source-id: 599bd43e5c3f717e2352f51e1bfac70bdbdbfc29