LockFreeVec
Summary:
(Extracted the diff from the following diff D45555710).

Introduce a new data structure: `LockFreeVec<T>` with characteristics:
* no atomics for `get(index)`
* `push_at(index, value)` which is safe, but fails if failed from multiple threads (e.g. outside of lock)

Use it in mapping from `DiceKey` to `DiceKeyErased` in the following diff D45555710.

# Name

This structure is not universally helpful. However, developing it in `dice` crate is painful, because `dice` compilation is very slow. So I placed it in `shed`. Alternatively, it can live somewhere like `dice/dice_core_impl`.

Naming is also hard. It says "lock-free" but it is actually wait-free.

It says `Vec`, but it does not store data in linear memory. Common with `Vec` is that `get` is `O(1)`. And there's `push` operation.

Reviewed By: JakobDegen

Differential Revision: D45638486

fbshipit-source-id: 73072a994fe934aaa74a498715710a3224b45493