default_only is a property of attribute, not of AttrType
Summary:
`AttrType` is a type of value in the attribute. And `DefaultOnly` is not a type.

Practically, we store redundant information in attr literal, which is dep attr type. Which occupies space.

While traversing attributes (to resolve or to collect dependencies) we rely on the fact that dep attr contains dep attr type.

Instead we can take it from attr type. Following diff D45422141 does that.

But to make it possible, we should store proper attr type for given attribute. This diff achieves that.

Reviewed By: ndmitchell

Differential Revision: D42833560

fbshipit-source-id: dd2e4b07f76acc7b62292d46b7e77097c0f7183c