buck2: span tracker: allow tracking roots mapped to DICE keys
Summary:
We want to show *some* form of estimate of work remaining to do in a build. We
don't get spans for those (and we really can't because they might actually be
DICE keys waiting on checking deps).

So, what we're going to try and do is just the number of pending DICE keys. The
problem with doing this is double-counting, since we might have some spans we
claim are roots and then we'll count them again because they are showing in
DICE.

So for each span we track as a root, let's also track what kind of DICE key it
represents (if any), and then we can just subtract those, and get a
reasonably-looking count.

Reviewed By: stepancheg

Differential Revision: D45524798

fbshipit-source-id: fb85032f84ce9e644dc3fdd1ed221517a9643c58