buck2: active_commands: use `span_tracker::Roots`
Summary:
So that I can reuse the same logic we use for tracking in the CLI.

It really sucks to have this duplicated, and I'm trying to reuse as much as I
can without pulling it in the entire EventObserver, which is fairly large
amount of work to be doing at runtime (actually doing it is easy, it's the perf
that's a concern).

This is, however, hopefully temporary, since in T149596463 we *will* actually
use EventObserver here (but we'll stop having it in the CLI side).

Differential Revision: D45563244

fbshipit-source-id: 5fef8de3876001c076c883b6e864c45bbb0c14d9