buck2: active_commands: put the state behind a mutex
Summary:
I want to put a little more data there and it becomes a bit tricky to fit
everything into a single AtomicU64.

This is mostly written to by a single thread, then periodically pulled by
different threads, so this should be rather fast, since the atomic will mostly
be local to the writer (pump_events) thread.

Reviewed By: stepancheg

Differential Revision: D45534711

fbshipit-source-id: 68f0587563dd98392a5876aa1b29bafa6277e599