buck2: buck2_server: switch to parking_lot
Summary:
I want to use it in the next diff here (because it's generally faster and I
want to use it in a mildly-performance-sensitive context), but also it's
arguably a better fit since this entire module does not want poisoning, which
is largely the reason to use the std Mutex vs. parking lot.

Reviewed By: stepancheg

Differential Revision: D45534710

fbshipit-source-id: 959c05aaf454816ac02d5c4705f665bc77f7187d