buck2: timed_list: don't try to color or add a time on the "X more" line
Summary:
This doesn't really make that much sense. We're coloring the "and X more" line
using the details of the span that would have gone there, but that line
represents potentially many other spans, so that's not really meaningful.

This makes it simpler, makes our console a little less red, and makes my next
diff easier.

Reviewed By: stepancheg

Differential Revision: D45524795

fbshipit-source-id: 08ff8b3f570e7f0ff6672719280fe36a50594202