buck2: events: only query hostname once
Summary:
Because it's not really something that changes after we started the daemon
(well, it could, but definitely not in the way we use this internally, and
this used to only be used for internal builds).

Also, while at it, don't gate this behind being a non-OSS build, because
it works fine in OSS builds.

The reason I'm doing this is because I'm going to query this a bit more in
the next diff.

Reviewed By: lmvasquezg

Differential Revision: D45688570

fbshipit-source-id: ff7b841397c66d8ab3b2a22aada065764b46e520