Collect reverse deps in rdeps using OrderedSet
Summary:
In prior D45289042, it was proposed that the new `rdeps()` implementation collect parents (i.e., back-edges in the dependency graph) in an `OrderedSet`.

This diff maps every dep to its parents as `NodeRef` -> `OrderedSet<NodeRef>`.

Previously, this mapping was `NodeRef` -> `Vec<NodeRef>`.

Thanks to iguridi for noticing the non-determinism of `HashMap` was bad and so `OrderedSet` was needed.

Reviewed By: scottcao

Differential Revision: D45709519

fbshipit-source-id: d1ee7ae8bfcc1eca8f4b7fb48ec2b817b8f0ba44