Add link execution preference attribute
Summary:
When linking, we currently have arguments for `local_only` and `prefer_local` to be set on the link action.

However, there are situations where we'd like to link remotely (and avoid downloading files to the user's computer during hybrid execution).

To express this, introduce link execution preference, which can be one of the following values:
- any (no preference)
- local_only
- local
- remote

In the Apple world, we also want to be able to set a execution preference based upon ***what*** is being linked. In order to support this case, allow setting either,
- an enum of the preference types
- a target that provides `LinkExecutionPreferenceDeterminatorInfo`, which contains a function that takes a list of targets to link and returns a preference

The default value for the attribute is None, allowing us to designate when no value was set, and in the next diff, use this to determine when to fall back on the toolchain's default preference.

Reviewed By: blackm00n

Differential Revision: D45673511

fbshipit-source-id: c7a182c0703d7f5382dcc8d3fd1a68683b4587ce