Voltron support for cxx libraries that only support some cpus
Summary:
For Voltron we construct a "target graph" using the information that is bubbled up in `AndroidPackageableInfo`. Previously, we only use the `AndroidPackageableInfo` for the "primary CPU" (the one that is used when building non-native libraries). This means that we can miss some native libraries that are only supported for a CPU that is not the primary CPU.

Instead, we construct the Voltron "target graph" using the `AndroidPackageableInfo` from all CPUs.

Reviewed By: navidqar

Differential Revision: D45726549

fbshipit-source-id: 829a592bd9dcbac1fbb3e8f7a0b2c35b37fd0bf3