buck2: dice: add a mechanism to get notified of key "activations"
Summary:
This allows the caller to observe what keys are being evaluated or
reused (hence: "activated") during a transaction in DICE.

The end goal is to use this for critical path calculations to fix the gaps we
currently have in them when early cutoff kicks in, which causes our critical
path calculation to get zero signal. This sometimes leads to stupid situations
like a 20 minute build with a 2 minute critical path (see the attached task).

I would actually like this to be per-DICE-version and not per-transaction, so
that it works properly with concurrent commands too, but that's a harder change
with smaller payoff (and telemetry for builds with concurrent commands is
currently totally busted anyway), so let's do this first.

The API is pretty heavily pattern-matched off of the "user cycle detector"'s
API for consistency.

Note that this currently (like the user cycle detector) only works in legacy
DICE, I'll add this in modern as well in time.

Reviewed By: bobyangyf

Differential Revision: D45602534

fbshipit-source-id: 9f763a68c91076c7a591cb9326136e716ab15b05