Cancel tasks when all waiters are dropped
Summary:
Cancel tasks when there's no one waiting for it anymore.

This only cancels when the last waiter gets removed.
Spawning a task (which would yet to have waiters) does not cancel the task since a waiter would immediately attach to this after creation in modern dice code.
We only care to cancel if the last dependent stops caring about it

Reviewed By: stepancheg

Differential Revision: D45558568

fbshipit-source-id: 1d610f3a85ef5f0a8e55943ebe0fdeb71b67f0c4