Make a separate type for base transaction in modern dice
Summary:
The base context that starts computations should be a different type.

The base context will hold the guard to keep cache alive
D45634599

Each keys compute will be subject to cancellation based on when the base context is dropped.

Also the base context will not hold parent_key, but percomputectx does

Reviewed By: stepancheg

Differential Revision: D45596860

fbshipit-source-id: 96694a5df3b77dd1ce2c00ed5c6a1ee802d0679b