Add cancel to DiceTask
Summary:
so we can wait request for cancellations directly on the dice task.

We already do trigger cancellations based on waker counts. This only deals with when keys stop depending on each other, or cancelling our dependency checks. There's another scenario where the top level computation context is dropped, for which we would care to wait for all tasks in dice to be terminated and cancelled.
This method will be used to implement the draining of tasks for that scenario.

Reviewed By: stepancheg

Differential Revision: D45596862

fbshipit-source-id: 66f2699f959a93a156e06b0cc62186807e6fb9fe