Implement ctx.actions.promise_artifact()
Summary:
This implements a function `ctx.actions.promise_artifact()` that
supports converting a promise to an artifact. The produced artifact can be
mostly used like any other artifact except that details about it (primarily
about its path, like extension/basename/etc) cannot be called until the
promise is resolved.

This is implemented by basically having the artifact hold a
`Arc<OnceCell<Artifact>>` that gets resolved at the end of analysis (when the
promises are resolved). There are places where we capture artifacts early on
(i.e. convert starlark artifacts to their rust ArtifactGroup form) and so
this requires that ArtifactGroup have a case to represent these
PromiseArtifacts.

I tried several other approaches to support this that wouldn't require a
representation in ArtifactGroup, but there's a lot of code that extracts artifacts
out of commandlinearglikes (and starlarkartifactlike) before freezing
(associated artifacts is a big one, but also a lot of actions, particularly
symlink_dir also do this extensively). I think ideally we'd change commandlinearglike
to have different apis for the frozen and unfrozen form (and the unfrozen would be
quite limited) and I tried that but I didn't see a route to accomplishing it
without extensive changes.

I think we could lift some of the limitations on the promise artifact by making the
user assert things when they convert it to an artifact. For example, it currently
requires that the promise resolve to a build artifact, but we could have the user
specify whether it will resolve to build or source and then validate that when its
resolved and we could similarly handle some information about the path. For now I
wanted to make this available so that that work could be guided by which of the
limitations people find objectionable when trying it out.

I think a nice improvement would be to extend this to support asserting that a promise
will resolve to a cmd_args, but I think that without that significant refactoring to
make it so that only the frozen form needs to be complete it will be really difficult
to support that.

Reviewed By: krallin

Differential Revision: D45684000

fbshipit-source-id: 209c660c4f25f53777bf6f059a733187ecebc1da