Determine whether to link locally based upon whether previous links were local
Summary:
The link ouput for iOS apps are large. When linked locally, in order for dependent (say dylibB) to link, the locally linked dylib (dylibA) must be uploaded, dylibB linked remotely, and then dylibB downloaded.

Instead of incurring the network cost to upload, if a user is linking locally, perform all subsequent links locally.

To achieve this, we need to make several changes:
1. Return through the linking callstack, the newly defined `LinkExecutionPreferenceInfo` which contains the execution preference utilized for the link
2. Provide it as an output of the target
3. When calling `preference_for_links`, accumulate all the `LinkExecutionPreferenceInfo` of deps, and pass it as an argument (in addition to the links)
4. In the `apple_selective_debugging`, utilize this information to determine whether to link locally or not.

Reviewed By: blackm00n

Differential Revision: D45673507

fbshipit-source-id: 015c52e09b8781e8e434a1ce9312a69adf71702b