Determine link_execution_preference based on whether we're also materialize debug symbols
Summary:
When selective debugging is enabled, we want to determine whether to link locally or remotely based upon whether we're also going to need the object files for debug data.

Thus, use the selection criteria of selective debugging to find all the targets being linked, and if more than 20% of the links will have debug data also downloaded, link locally. Otherwise link remotely.

The 20% is arbitrary, and right now we make a determination on whether to link locally or remotely only. We could also set a preference of `any` which would run on hybrid. We'll likely need to tinker with these options to find the best setup for users. For now, this gets the basic integration in place.

Reviewed By: blackm00n

Differential Revision: D45673513

fbshipit-source-id: 7d47f5932a1080f831fb9904b27f182a1902081e