buck2: uploader: in OSS, be happy with any TTL > 0
Summary:
In OSS, we don't actually TTLs back from RE (unlike internally when we do).
Our client there just makes up a TTL of 60 seconds.

The problem is that this is less than 600 seconds we want here, so we'll
reupload things even if they aren't missing.

Accordingly, let's just not expect any particular TTL, and be content with
anything nonzero (where zero would indicate the blob is not there).

An alternative would be to return a higher made up TTL (higher than 600 seconds),
but we do actually cache this. There's unfortunately very little in the way of a spec
to tell you how long a find_missing result is valid for, but it seems fair to assume
60 seconds, so that's what we do.

Reviewed By: lmvasquezg

Differential Revision: D45811294

fbshipit-source-id: 3579da3e0315b7af1f930fb6b54579796e08782c