Integrate hyper-based http client with action impl + materializer
Summary:
Now that we have a functioning hyper-based http client, let's use it instead of the reqwest-based client.

This moves the `http_client()` fn which instantiates an http client depending on if this is OSS or not into `buck2_common`. It's expected that this should be a general-purpose function for other users of http clients to call and get the right kind of client for their use case.

Reviewed By: krallin

Differential Revision: D45456986

fbshipit-source-id: 9bebd8c351a48102fb33d72c36b1aac2a17cdc7b