http client: Create http client for materializer and action executor
Summary:
We need to create the new http client for two subsystems: materializers, and action executors.

For materializers, we can instantiate it directly on materializer creation.

For action executors, the best way is to pass this to dice and fetch it like other global-ish state.

Unfortunately, this means we're constructing the client in two places; it would be better if we could share client creation in one callsite somehow, but doing so requires putting it on `(Base)ServerCommandContext` which feels like not the right place.

Reviewed By: krallin

Differential Revision: D45647892

fbshipit-source-id: 4ea02bc5b68d79ce0f79a87e52d42c1ca1e07667