Revert D45419945: Multisect successfully blamed D45419945 for test or build failures
Summary:
This diff is reverting D45419945
D45419945:  [buck2] Enable multiprocessing support in inplace par" by asp2insp has been identified to be causing the following test or build failures:

Tests affected:
- [starlight/aienv/src/v2_compat/api/tests:test_app_api_v2 - test_run (starlight.aienv.src.v2_compat.api.tests.test_app_api_v2.TestAppApiV2)](https://www.internalfb.com/intern/test/562950042050290/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2056003
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: IanChilds

Differential Revision: D45839888

fbshipit-source-id: 8ee172c62fb9db4668c373624f38c31d2dab7ef4