Fix up error checking
Summary: This function needed quite a lot of attention. There were some errors that could panic (e.g. if you modify a dict to insert a non-string key) and if it did error, it didn't always say which key was at fault. Plus a lot of the artifact checking code was duplicated three times. Clean all that up.

Reviewed By: zsol

Differential Revision: D45861795

fbshipit-source-id: ab08b1ce81891b1009d0b381bcd51e6a013e86c2