Delete compute_many
Summary:
Prepare to change API to solve eager deps

The plan is to have change Dice to accept some kind of KeyLike/RequestToken that can be converted and polled on as futures in specific combined ways.
We'll no longer just receive user `Key`s as simple as we do now.

The current dice compute API's are insufficient.
No one really uses compute_many, so get rid of it for now to keep it simple as we'll only need to deal with one entry point to dice (`compute`), and we'll add the new variants later.

Reviewed By: stepancheg

Differential Revision: D45800933

fbshipit-source-id: fda3ec89f40aa0c969ec45c37c636cba92677853