Remove "Jobs" word from TimedList
Summary:
On one hand, everything buck2 does is kind of "jobs", on the other hand even internally we don't call "jobs" the thing described on that line.

It is clear from the context what is going on on that row, and "jobs" does not help.

This will also truncate less in narrow terminals.

Differential Revision: D45846416

fbshipit-source-id: add707ce5e3552bb6cea10a766b94f287a85b92e