Make sure future is dropped on completion
Summary:
Trying to solve the panics in the tasks attached.

Previously the cancelled future would say it is terminated before it drops future.

That future would only be dropped upon the next poll, so depending on the `pin` and poll stack, the future may hold onto objects beyond the lifetime for which it is considered "terminate".

I suspect this is causing the Arcs to be alive and causing panics in dice due to the race where the thread that observes and wait for termination will receive the signal that the future has terminated before the worker thread polls and drops the future, which would keep data structures alive in the future that shouldn't be alive if the future was terminated.
This way, we guarantee that the future drops all datastructures, then notify any other threads that the future has terminated, so other logic can safely assume the termination means the future is completely gone.

Reviewed By: krallin

Differential Revision: D45622039

fbshipit-source-id: 0d5ce31d2ea5d0a79e49fcc4a2b521dba70cc34d