Future isn't considered terminated until polled and dropped when cancelled
Summary:
if we consider ourselves terminated before the first poll, that would mean we consider the future terminated without ever dropping the future.

We could fix this by drop on cancel, but that requires the future being droppable from different threads, which is more complex.
Instead, since we should always be spawning, we just assume that we'll be polled once, and have that trigger cancellation

Reviewed By: krallin

Differential Revision: D45634523

fbshipit-source-id: caa1210c0dcc22c466c7c3073353b89aaa5f619e