Have the lambda for spawn execute immediately instead of on poll
Summary:
This causes a lot of caveats later since the lambda is delayed, where things get quite racey in modern dice

Simpler to have non-future code be not lazy, and only async code be lazy.
This is a nice guarantee and keeps code simpler to think about.

One example issue is that `DiceTask` creates a TaskHandle based on the CancellationContext. With this lazy, that handle would only be created once the future is actually polled.
If the future gets cancelled before poll, then that TaskHandle would never be created, and drop logic never be executed. Without this, we would have no ability to non-lazily do operations with the CancellationContext.
It's possible to get dice to work around this by adding extra drop guards, but that's complexity.
It's easier to reason about not having to worry about the laziness of future spawning.

Reviewed By: JakobDegen

Differential Revision: D45628795

fbshipit-source-id: cf09772e3de111a6bcc9ea69fdb80b81be6ae159