Make BaseComputationCtx keep the active version alive
Summary:
Futures from ctx are tied to lifetime of context, so to drop the context, futures must be dropped

Therefore, we make the BaseComputationCtx (which is the top level context) keep the active version alive. If this is dropped, then the active version is dropped, which indicates top level requests has been cancelled and we should trigger cancellation of all tasks anyways.
i.e D45634595

This makes it so that we don't depend on the entire dropping of all requests.

Reviewed By: stepancheg

Differential Revision: D45634599

fbshipit-source-id: 40effdd67d1345a4d7388c87d5bc523345bbe82a