Make temporary_spawn only receive reference
Summary:
temporary_spawn current gives access to the whole DiceTransaction, but given that it is callable from within a Key's compute, this is wrong and something that I wanted to remove since the beginning.
Now with the new spawns, separate BaseCtx vs PerComputeCtx, and solving the lifetime inference issue of async lambdas, this is finally possible.

`temporary_spawn` was only supposed to temporarily exist in our code base because of the unsoundness.

Unsoundness explained:
The `compute`'s ComputationCtx holds the same underlying data as the transaction. At the end of the `compute`, we expect the ComputationCtx to no longer have any active references, hence `finalize`ing it.
Meanwhile, `temporary_spawn` provides ownership to a Transaction (which is wrong because it also gives you access to update dice, which a compute should never do). By owning it, we disassociate the lifetime of the action ComputationCtx from the future that is spawned. This can cause problems since temporary_spawn can result in a future that holds onto the underlying data beyond what we expect, causing `finalize` to fail.

Reviewed By: krallin

Differential Revision: D45634598

fbshipit-source-id: 852116ba792d790baf38600fc7aac06785117b64